description (recommended to read)
For laboratory work #8, we had to write the code for dragging and dropping nodes. Since I couldn't write the code for dragging and dropping to each node, I wrote for the whole svg object. As the result, the dragging part works well, however, in dropping part, there always appears null.